+++ src/lib.rs
@@ -306,32 +306,56 @@ where
     }
 }
 
-/// Returns an iterator of all Gaussian primes *a* + *b*i
-/// where |a|,|b| ≤ `n`.
-pub fn get_g_primes(n: isize) -> impl Iterator<Item = GaussianInt<isize>> + 'static {
-    let mut primes: Vec<GaussianInt<_>> = vec![];
+/// Returns an iterator of all Gaussian integers *a* + *b*i
+/// where |*a*|,|*b*| ≤ `n`.
+pub fn get_g_ints(n: isize) -> impl Iterator<Item = GaussianInt<isize>> + 'static {
+    let mut integers: Vec<GaussianInt<_>> = vec![];
     for a in -n..=n {
         for b in -n..=n {
             let z = GaussianInt::new(a, b);
-            if z.is_gaussian_prime() {
-                primes.push(z);
-            }
+            integers.push(z);
+        }
+    }
+    integers.into_iter()
+}
+
+/// Returns an iterator of all Gaussian integers *a* + *b*i
+/// where *a* is positive (or zero) and |*b*| ≤ `n`.
+pub fn get_pos_g_ints(n: isize) -> impl Iterator<Item = GaussianInt<isize>> + 'static {
+    let mut pos_integers: Vec<GaussianInt<_>> = vec![];
+    for a in 0..=n {
+        for b in -n..=n {
+            let z = GaussianInt::new(a, b);
+            pos_integers.push(z);
+        }
+    }
+    pos_integers.into_iter()
+}
+
+/// Returns an iterator of all Gaussian primes *a* + *b*i
+/// where |a|,|b| ≤ `n`.
+pub fn get_g_primes(n: isize) -> impl Iterator<Item = GaussianInt<isize>> + 'static {
+    let set = get_g_ints(n);
+    let mut primes: Vec<GaussianInt<_>> = vec![];
+    for z in set {
+        if z.is_gaussian_prime() {
+            primes.push(z);
         }
     }
     primes.into_iter()
 }
 
-/// Returns an iterator of all Gaussian integers *a* + *b*i
-/// where |*a*|,|*b*| ≤ `n`.
-pub fn get_g_ints(n: isize) -> impl Iterator<Item = GaussianInt<isize>> + 'static {
-    let mut primes: Vec<GaussianInt<_>> = vec![];
-    for a in -n..=n {
-        for b in -n..=n {
-            let z = GaussianInt::new(a, b);
-            primes.push(z);
+/// Returns an iterator of all Gaussian primes *a* + *b*i
+/// where *a* is positive (or zero) and |*b*| ≤ `n`.
+pub fn get_pos_g_primes(n: isize) -> impl Iterator<Item = GaussianInt<isize>> + 'static {
+    let set = get_pos_g_ints(n);
+    let mut pos_primes: Vec<GaussianInt<_>> = vec![];
+    for z in set {
+        if z.is_gaussian_prime() {
+            pos_primes.push(z);
         }
     }
-    primes.into_iter()
+    pos_primes.into_iter()
 }
 
 impl<T: PrimInt + Integer> One for GaussianInt<T> {