+++ src/lib.rs
@@ -101,11 +101,13 @@ impl<T: PrimInt + Integer + Signed> GaussianInt<T> {
     /// # use gaussiant::GaussianInt;
     /// # fn main() {
     /// let z = GaussianInt::new(2, 7);
-    /// assert_eq!(z.norm(), GaussianInt::new(53, 0));
+    /// assert_eq!(z.norm(), 53);
     /// # }
     /// ```
-    pub fn norm(&self) -> Self {
-        *self * self.conj()
+    pub fn norm(&self) -> usize {
+        let a = self.0.re;
+        let b = self.0.im;
+        (T::pow(a, 2) + T::pow(b, 2)).to_usize().unwrap()
     }
 
     /// Returns `true` if `self` divides `other`.